Avoid tests failing after they are completed
Running the package tests with Go 1.16 it reports stack traces for the test
failures.

The pattern of these fixes is the same: ok being true now signifies that the
main control flow is expected to operate during that time.  That main flow
sets ok back to false once it is done, and the test won't fail if the late
goroutine encounters that state.