Fix tests to not call t.Fatal from a goroutine
The "too late!" version of this call was being made *after* the
corresponding test had completed, causing all manner of sadness from the
testing package.

This refactors the tests to just measure the duration of the various
kinds of sleep, and assert that the duration is within a good range.