Fix race condition in Mock.AfterFunc()
Previously, AfterFunc() created a timer using Timer() and then added a
function callback to the timer. This created a data race because the
timer could fire between when it is created by Timer() and when the
function callback is set.

Instead, this change created the Timer{} struct and adds it to the
Mock's list of timers inline. This ensures that the function callback is
set before the Timer is able to fire.